+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Changed
 Removed
 
 Fixed
+- Protocol.stamp() allows one_tip=True when steps use a `mix_vol` greater than "31:microliter" even if transferred volumes are not all greater than "31:microliter"
 
 ## v2.5.0 - 2015-10-12
 ---